requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0